patch SdrPageWindow in SdXImpressDocument::paintTile()
This is based on commit 793fc965195c7f9e51d7f2685eac433c2139a62b
amd I don't quite understand everything that happens, because
the drawing layer classes have all kinds of comments except those
that would explain the purpose or usage of the classes. My
understanding is that drawing layer keeps all kinds of state
in Sdr* classes and when drawing it tries to find those based
on the OutputDevice. But since tiled drawing uses its own
VirtualDevice for the drawing, the search fails and a temporary state
is used that's always setup up and cleared again. Which is wasteful,
and not having the state around also means that e.g. when clicking
a checkbox for impress master background, with idle painting
disabled there's nothing to react to the event and cause a tile
repaint.

Change-Id: Ic9036680c628d67615a82dcc4f585ab827b91525
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121780
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>