Clarify that css::sdbc::XClob::getSubString takes 1-based pos
... and fix connectivity::firebird::Clob::getSubString.

XClob::getSubString is modelled after JDBC counterpart, that also
takes 1-based position argument. This is also described in multiple
places in our code:

- OPreparedStatement::setClob
  in connectivity/source/drivers/firebird/PreparedStatement.cxx
- OSingleSelectQueryComposer::setConditionByColumn
  in dbaccess/source/core/api/SingleSelectQueryComposer.cxx
- ORowSetValue::getString
  in connectivity/source/commontools/FValue.cxx

However in some places 0-based value was used (fixed here).
To clarify, the mention that the pos argument is 1-based is put
to the corresponding IDL file.

Also the code in connectivity::firebird::Clob::getSubString had
multiple issues:

- no checks of arguments;
- possibility to throw "nPosition out of range" when just-read
  segment has enough data;
- wrong start position in case when nPosition is not aligned to
  segment boundary.

This change fixes these, and simplifies the implementation.

Change-Id: I119a62dd7f02c9569ce36395ed8cc1a98c931fcf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121884
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>