tdf#115547: Fix DB path handling
1. The code used wrong procedure to convert file URLs to local paths.
It assumed that it's enough to just strip the leading 'file://', which
is only sometimes true on Unix-like systems; on Windows, this converts
a valid 'file:///C:/path/file.ext' to '/C:/path/file.ext', where the
leading slash is then treated as a network path, resulting in errors.

2. It is incorrect to assume the same length for UTF-16 and UTF-8
encoded URLs coming from untrusted source (like user file). It may
contain non-ASCII characters (be an IRL), and then the assumption
would fail.

Change-Id: Ie2ea6c8cb9a690975db956fa025bf926a8010984
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121885
Reviewed-by: Lionel Mamane <lionel@mamane.lu>
Tested-by: Jenkins