sw: fix undo manager de-registration from editeng on shutdown
editeng/ asserts in the ImpEditEngine dtor that the undo manager it owns
is indeed an EditUndoManager, and not a subclass. This means that if a
subclass is set as the undo manager of editeng, then SetUndoManager()
should be called in pairs: once to set it and once to unset it. The
unset should happen before destroying the EditEngine.

This normally works because SwView either has an active text edit and
then its dtor calls SdrEndTextEdit() or it doesn't have an active text
edit.

The broken case was when the text edit was already ended, but the draw
view still had a text edit outliner, which has an EditEngine, which
knows the sw undo manager. The product build deleted the sw undo manager
when deleting the EditEngine, and then later when the SwExtTextInput
dtor wanted to access the sw undo manager, it crashes due to
use-after-free.

Fix the problem by explicitly disposing the undo manager of the draw
view in the dtor of SwView.

Also fix a couple of more places where an SdrObject* is returned and we
didn't check if the result is a nullptr.

Caught by the loolstress tool in online.git:

cp test/data/hello-world.odt /tmp/test.odt
./loolstress ws://localhost:9980 /tmp/test.odt test/traces/writer-hello-shape.txt /tmp/test.odt test/traces/writer-hello-shape.txt /tmp/test.odt test/traces/writer-hello-shape.txt /tmp/test.odt test/traces/writer-mash-text-table.txt /tmp/test.odt test/traces/writer-mash-text-table.txt

Change-Id: Ib838b2adf900b4f3bec63d2d62d432327bc0c6c4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122086
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins