optimize SwRegionRects::Compress()
This function gets called on modifying documents, and it shows up
noticeably e.g. when profiling editing in Online.
The code repeatedly restarts the whole process on most changes,
but in practice this seems to only lead to repeated checks with
the same outcome. So finish the whole compression and try again
only at the end, if deemed needed. And even there I'm not sure
if it's actually ever needed, since I couldn't trigger such a case
in practice.

Change-Id: I15b0a83c6865bad21fe60a7e9ca0b87ceea5b489
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122156
Tested-by: Luboš Luňák <l.lunak@collabora.com>
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>