wina11y: Drop AccObject::m_bShouldDestroy
The getter, 'AccObject::ifShouldDestroy',
was unused, so drop it.

The only place where the 'm_bShouldDestroy' member
was actually used was in 'AccObject::NotifyDestroy',
where it was assigned a value from a method param that
was then passed further right away, so there's no need to
store the value at all.

Change-Id: Id5688e9742892e345b143a8eebeed5099696a0f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122342
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>