Fix splitting of "permission-for-group/-user:" strings
...that had been introduced in 5628bf1c429c058dc46a30e21576cf489a7ad8bc
"tdf#66398 Parse and output permissions for DOCX using bookmarks":

For one, in DocxAttributeOutput::DoWritePermissionTagStart permissionIdAndName
was left empty in the permission-for-user case.  This had already been addressed
for DocxAttributeOutput::DoWritePermissionTagEnd in
0a30aed08121a0077a37183c0ff6d211abe010ca "permissionIdAndName is empty in
permission-for-user case".  But from the calling code it looks like in both
functions that permission will always start with either "permission-for-group:"
or "permission-for-user:", so assert that rather than silently doing nothing if
the assumption is wrong.

For another, in both functions the code didn't take any precaution if
permissionIdAndName didn't contain a colon (which would cause undefined behavior
when OUString::copy is called with a negative count argument).  Again, from the
calling code assume that permissionIdAndName will always contain a colon, and
assert that.

Change-Id: Ia7432413c138d47efbe8aec76f14c561ac241fb3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122387
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>