qt5 a11y: Implement QAccessibleTableCellInterface
Let 'Qt5AccessibleWidget' derive from 'QAccessibleTableCellInterface'
and implement the most important methods.
This is e.g. one of the steps needed to make Orca announce
the focused cell in Calc.

Since there's no specific XInterface for table cells
make use of the fact that a table cell's parent is a table
and query information from there, similar to how the following
commit does for winaccessibility:

    commit dae4088fc8188940bcddd2d20982300dda608d67
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Sep 8 14:37:53 2021 +0100

        tdf#100086 tdf#124832 wina11y: Implement IAccessibleTableCell

Change-Id: I160bc04f3e4fcf7b77723540aba6945b8fdf36ae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122395
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>