memory leak on loading dbase files
see on loading qadevOOo/testdocs/BasicDB/DBTest.dbf

a problem apparently since...

commit c43937b07de9644dad33b04f8054fdc7d04d5794
Date:   Tue May 8 12:42:55 2018 +0200

    loplugin:useuniqueptr in ORowSet

OTableContainer inherits from OCollection which has a unique and
nonstandard acquire/release which forwards to a parent
cppu::OWeakObject. So when it would appear to a casual observer, who
assumes that it has an internal refcount, that its refcount would drop
to zero and then get auto-deleted, nothing of the kind occurs.

Change-Id: Ib104415ccc13c6ed9c06f975857fcd4413e2e05d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122519
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>