leaks of OCollection in dbase import
presumably since...

commit fd3439968c7594f9aa2151fb848729db3df4e9a7
Date:   Mon Jan 29 13:28:44 2018 +0200

    loplugin:useuniqueptr in OTable

OCollection has a unique and nonstandard acquire/release which forwards
to a parent cppu::OWeakObject. So when it would appear to a casual
observer, who assumes that it has an internal refcount, that its
refcount would drop to zero and then get auto-deleted, nothing of the
kind occurs.

Change-Id: Ic292663a61b4d7dbf7636be7ad4c00aed0b5a7fa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122528
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>