tdf#45525 tdf#64242 sw row optimize: get correct row height
Optimizing is trying to find the actual height of each row
and adding them all up together to find the total height.
But without this fix, the height of the previous rows
was never being reset, so the later rows were potentially
reporting a larger height than they should have.

Change-Id: I789edc0ca700d9b1448832fa9720ec7ec3b16223
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122572
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>