tdf#88163 Fix font size for placeable wmf files
The problems in tdf#88163 can be categorized into two parts, as
described in 1aab5506a65234c38a2b4f14a1bd1155007d63c6:

First, the problem with the wmf files without the placeable header.
Second, the problem with the wmf files with the placeable header.

The above mentioned patch fixed the first part, and this patch fixes
the second part.

The problem was that upon seeing 22-byte placeable header, the
records related to the size of the wmf like META_SETWINDOWORG (0x20b)
and META_SETWINDOWEXT (0x20c) and others were ignored. These records
were read in WmfReader::GetPlaceableBound() for the wmf files without
placeable header. Adding this method for the wmf files with placeable
header fixed the wrong calculation of bounds, which previously lead
to wrong size of text.

It should be noted that the scale in the placeable header is used,
but the bounds are ignored for now.

A new test named testTdf88163PlaceableWmf() is added that can be
checked with:

    make CPPUNIT_TEST_NAME="testTdf88163PlaceableWmf" -sr \
        CppunitTest_emfio_wmf

Change-Id: I820c2e5922972cb5d555d98ef70c7581cd9f02d5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122095
Tested-by: Jenkins
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>