Resolves: tdf#144740 Fix broken compareByValue() query, tdf#142910 tdf#144253
Fix regression of a series of commits that, intended for filter
queries, unconditionally round numeric values as shown under
ScTable::ValidQuery() and compareByValue() without having taken
into account that the same query and compare functions are used by
the interpreter for all functions that use query criteria,
possibly delivering completely wrong results including in
backports to 7.2.0.0

    commit 6b0f0ac92a896dbe83c6f9fced883f02ae6f15e8
    CommitDate: Wed Jul 7 17:44:46 2021 +0200

        tdf#142910 sc filter: fix "greater than" or "smaller than" etc

    commit 8abb7893b4312df5d2183b2aaa8eb6fb1737f616
    CommitDate: Tue Sep 21 11:06:35 2021 +0200

        tdf#144253 tdf#144324 sc filter: use formatted values in filters

Several related and intertwined commits in filter context make
assumptions about these queries always being executed rounded, so
the only clean solution is to make that depend on the
ScQueryEntry::Item being passed. Its mbRoundForFilter value is set
to true for all items of all queries executed via ScTable::Query()
and ScTable::GetFilteredFilterEntries(). It might be not all are
necessary (or some even still harmful?) and unnecessarily
obtaining number formats and calling RoundValueAsShown() is still
a bottle neck for those, but that should be addressed and reworked
independently. The important part is calculations work as before.

Also, moved obtaining number formats for calling RoundValueAsShown()
into logic that calls them only if necessary.

Note the TODO in compareByValue() about suspicious rounding of
rItem.mfVal in filter context that is to be addressed.

Change-Id: Ieb178ad1ea15a635caeb1ba698c2f4b7ad676d57
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122729
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins