tdf#141633 use OutputDevice::DrawText for multiline text too
m_bFastPaint is only false for the multiline text, where it then tries
to use a textview widget to render the text onto the OutputDevice while
for simpleline text it just uses OutputDevice::DrawText.

This isn't working particuarly well anymore for print output and
DrawText can be used with flags of DrawTextFlags::Top |
DrawTextFlags::MultiLine | DrawTextFlags::WordBreak to render text
equivalently to the multiple line textview which seems a saner approach.

Change-Id: I3c96f837aaeba82ebf5956ab2ba35210c978f255
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122953
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>