ucb: webdav-curl: try to prevent acquiring a lock that we already have
This was done for webdav-neon in commit
73919cccceecd301f9f7a4c404f02e379d8f05cc but without checking that the
requested lock isn't "stronger" than the one that was already acquired.

This is currently not a problem because there is only one caller of
LOCK() and it passes a hard-coded ucb::Lock argument, but not sure if
that would change in the future, so try to check for it.

Change-Id: I7a703197208979c330212288c323877a326dd7de
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123229
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>