vcl: fix issue in testErase()
Funnily enough, this *could* still be a valid test because
SetOutputSizePixel() acutally triggers an Erase(). But that's not really
the point, the test is brittle because we possibly may remove the Erase
from that function one day.

This sets the actions we test to the *second*, explicit Erase we are
actually testing.

Change-Id: Ib0bb4693a90ad2a4756f93a8933db95d337950c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123236
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>