Converting pointer to reference in svgwriter
Previously, pElementId (old name) was either nullptr or a pointer to
OUString created from sPlaceholderTag literal.  In the methods that
consumed it, it was checked for being nullptr. Now, this is checked
with if the aElementId (new name) is empty or not.
If the variable is an empty string, it implies that it should not
be used.

Change-Id: I560fd479f428341c834ee99b0f8a34e95862e3e0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123439
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>