tdf#96499 sc htmlimport: fix broken CSSHandler so that...
...ScHTMLStyles can handle those MSO-Number-Formats again.

The CSSHandler was added by the following commit, after which the
sc htmlimport can parse the CSS stylesheets and set number formats
of table cells accordingly:

commit 414d16514654b7de2164cac2a982f4a47c7ff39c
Author: Kohei Yoshida
Date:   Thu Jul 28 00:46:55 2011 -0400
    Parse CSS in the <style> content and set number formats to cells.

However, this was broken since "selector_name" does not no
longer functions in the handler since upgrade to orcus 0.9:

commit 1cdf764b091d5173a5e8be0aa60ec2ad7649305f
Author: David Tardon
Date:   Thu Jun 4 16:09:48 2015 +0200
    update to liborcus 0.9.1

And further the selector_name was then removed by the
following commitd due to loplugin:unreffun checks:

commit 55eb7d68b825b8b4c980a624e2bb4d6572398676
Author: Miklos Vajna
Date:   Fri Jun 5 12:29:36 2015 +0200
    sc: loplugin:unreffun

commit 1ffd86863800b8fc88beb153851ff0479a12f05f
Author: Miklos Vajna
Date:   Fri Jun 5 12:32:10 2015 +0200
    sc: another loplugin:unreffun

And then this was further broken by the following commit, in which
it removed the mrStyles of type "ScHTMLStyles" from the CSSHandler
class:

commit 631f9ef5b7bfba255ce5d47a9489e5a427ded30f
Author: Noel Grandin
Date:   Tue Sep 18 09:57:26 2018 +0200
    loplugin:unusedfields improve search for unused collection fields

This patch partially reverted those code, revised the
CSSHandler member functions to make it work. More handler
functions need to be implemented.

Note that this bug is still not resoved yet because the orcus
css parser does not accept css strings to include unquoted
non-ascii chars as property values. This will be addressed in
a separate patch.

Change-Id: I8883f5a9362d9476dbfa9a2eed283c6f5c18305d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123715
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>