Resolves: tdf#145169 only position treeview cursor at the first entry
when the view is filled, don't actually select the entry.

With these single selection trees the default is to select when the
cursor is positioned, so move the unselect to after the cursor
positioning to achieve this.

Now nothing is considered "selected" in the view unless the user
explicitly does that so the contents of the entry are used by default
on save/load because nothing is selected in the treeview unless the
user does that.

Change-Id: I2a740616205bab35614f85f62e8240cd11df90ad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123754
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>