Related: tdf#145173 Ensure known supported system and default document locale
For an unsupported locale like {en-IL} that could had ended up as
language-only {en} if the aImplIsoLangEntries map contains such
entry as obtained through
MsLangId::convertUnxByteStringToLanguage() and
Conversion::convertIsoNamesToLanguage(). For the system locale in
SvtSysLocale a proper fallback is used but the default document
language was propagated as is

Other places evaluating MsLangId::getSystemLanguage() are affected
as well, those probably will have to be replaced by a call to
MsLangId::getRealLanguage(LANGUAGE_SYSTEM) or a newly to be
introduced MsLangId::getConfiguredSystemLanguage() and
MsLangId::getSystemLanguage() be made private or accessible only
by LanguageTag.

Change-Id: I87eb9456d5b4ea8d64b0c41fec6bd2739256fb56
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123756
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins